pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rulescope"
version = "1.0.0"
description = "Governance engine for Sigma and detection catalogs — semantic diff, maintainer hotspots, explain mode, PR regression gates, ATT&CK Navigator export, and plugin analyzers"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Jean Koumou", email = "jeank@eliesec.fr" }
]
keywords = ["sigma", "detection", "quality", "siem", "blue-team", "detection-as-code", "governance"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Information Technology",
  "Topic :: Security",
  "Topic :: Software Development :: Quality Assurance",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.7.1",
  "PyYAML>=6.0.1",
  "Jinja2>=3.1.4",
  "rapidfuzz>=3.9.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.9",
  "build>=1.2.2",
  "twine>=5.1.1",
]

[project.scripts]
rulescope = "rulescope.cli:main"

[project.urls]
Homepage = "https://github.com/Kjean13/rulescope"
Repository = "https://github.com/Kjean13/rulescope"
Documentation = "https://github.com/Kjean13/rulescope#readme"
Issues = "https://github.com/Kjean13/rulescope/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"rulescope" = ["py.typed"]
"rulescope.reporters" = ["templates/*.html"]

[tool.setuptools.packages.find]
include = ["rulescope*"]
exclude = ["tests*", "docs*", "datasets*", "examples*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "UP"]
ignore = ["E501", "E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=rulescope --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
branch = true
source = ["rulescope"]
omit = ["rulescope/watcher.py"]

[tool.coverage.report]
show_missing = true
skip_covered = false
